Added rdParam in initialization

+++ User/inc/main.h
@@ -4,7 +4,7 @@
 #include "ini.h"
 #include "it.h"
 
-uint32_t RD = 0;
+uint16_t RD = 0;
 uint8_t Flag_IRQ = 0;
 uint32_t n = 0;
